FAISS builds with Windows

+++ src/3rd_party/faiss/Index.h
@@ -67,7 +67,7 @@ struct Index {
     float metric_arg;     ///< argument of the metric type
 
     explicit Index (idx_t d = 0, MetricType metric = METRIC_L2):
-                    d(d),
+                    d((int)d),
                     ntotal(0),
                     verbose(false),
                     is_trained(true),
